Exploring the super-sample issue again
It seems that, because the in/egress are so short (15 min), the transit is quite distorted and fitting the data requires a very high super-sample factor (100 seems to give a model convergence to 2 ppm).

Of course, such a high super-sample factor makes the model run very slowly when the data are binned to 1-min, so as a compromise, I'm binning the data to  30-min bins. Additional analysis may change that approach, though.