Still trying to understand transit residuals
Still the residuals between my model fit to the transit and the actual transit data for Kepler-76 are very large (~100 ppm) and largest during in/egress, which suggests my model isn't properly capturing something.

In this version of the Analysis_of_Kepler76b.ipynb, I've experimented with subtract out the EV+ signals from the transit portion of the binned light curve. Even though the EV+ signal fit looks good, the best-fit transit residuals are large, which indicates that the problem lies in the transit model itself, NOT in the fact that I wasn't account properly for the EV+ signals.