Experimenting with BEER_curve
BEER_curve seems to do a fine job with fitting the data, giving <= 10 ppm residuals if I let it fit everything (except LDCs). And it only takes about a minute to fit.

The one problem now is that the eclipse still seems to have big residuals. Not clear what's going on there.