Need to fit BEER and eclipses at the same time
The enormous apparent variations I was seeing in eclipses are probably just due to improper removal of systematics/BEER signals. The deepest eclipse so far looks like nonsense (eclipse-fit_300ppm-deep_SNRgt5.png).

Instead, it makes more sense to fit the BEER signals for each orbit (or collection of orbits), subtract those signals, and THEN fit the eclipse.

So I'm going to re-tool the analysis to do just that.